data_pipeline/requirements.txt
# Data Pipeline Dependencies

# Core
networkx>=3.0
numpy>=1.24.0
scipy>=1.10.0

# APIs
requests>=2.31.0
httpx>=0.24.0  # For async API calls

# Clustering
python-louvain>=0.16

# Layout (CPU)
# fa2>=0.3.5  # Has Python 3.10 compatibility issues, using NetworkX spring_layout fallback

# Layout (GPU - optional)
# cudf-cu11>=23.0
# cugraph-cu11>=23.0
# cupy-cuda11x>=12.0

# ML/Embeddings
torch>=2.0.0
transformers>=4.30.0
scikit-learn>=1.3.0

# Search (optional)
# faiss-cpu>=1.7.4
# faiss-gpu>=1.7.4

# Database
psycopg2-binary>=2.9.0

# CLI
click>=8.1.0
tqdm>=4.65.0

# Config
pydantic>=2.0.0

# API Server
uvicorn>=0.23.0
fastapi>=0.100.0